--- a/OggVorbisEncoder.Benchmarks/OggVorbisEncoder.Benchmarks.csproj
+++ b/OggVorbisEncoder.Benchmarks/OggVorbisEncoder.Benchmarks.csproj
@@ -1,0 +1,22 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="BenchmarkDotNet" Version="0.12.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="../OggVorbisEncoder/OggVorbisEncoder.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Update="unencoded-30s.raw">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+  </ItemGroup>
+  
+</Project>