--- a/OggVorbisEncoder/Lookups/FloorLookup.cs
+++ b/OggVorbisEncoder/Lookups/FloorLookup.cs
@@ -101,7 +101,7 @@ namespace OggVorbisEncoder.Lookups
             }
         }
 
-        public int[] Fit(OffsetArray<float> logmdct, float[] logmask)
+        public int[] Fit(in Span<float> logmdct, float[] logmask)
         {
             var n = _n;
 
@@ -268,7 +268,7 @@ namespace OggVorbisEncoder.Lookups
             return null;
         }
 
-        private bool InspectError(int x0, int x1, int y0, int y1, float[] mask, OffsetArray<float> mdct)
+        private bool InspectError(int x0, int x1, int y0, int y1, float[] mask, in Span<float> mdct)
         {
             var dy = y1 - y0;
             var adx = x1 - x0;
@@ -427,7 +427,7 @@ namespace OggVorbisEncoder.Lookups
             return (a[pos] + b[pos]) >> 1;
         }
 
-        private int AccumulateFit(float[] flr, OffsetArray<float> mdct, int x0, int x1, ref FitAccumulation fits, int n)
+        private int AccumulateFit(float[] flr, in Span<float> mdct, int x0, int x1, ref FitAccumulation fits, int n)
         {
             int xa = 0, ya = 0, x2a = 0, xya = 0, na = 0, xb = 0, yb = 0, x2b = 0, xyb = 0, nb = 0;