--- a/OggVorbisEncoder/Lookups/ResidueLookup.cs
+++ b/OggVorbisEncoder/Lookups/ResidueLookup.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using OggVorbisEncoder.Setup;
 
 namespace OggVorbisEncoder.Lookups
@@ -165,7 +164,7 @@ namespace OggVorbisEncoder.Lookups
             var index = 0;
 
             // assumes integer/centered encoder codebook maptype 1 no more than dim 8 
-            var p = new int[8];
+            Span<int> p = stackalloc int[8];
             if (book.Delta != 1)
                 for (i = 0, o = book.Dimensions; i < book.Dimensions; i++)
                 {