… OggVorbisEncoder/OggVorbisEncoder.csproj
… OggVorbisEncoder/ProcessingState.cs