@SteveLillis 2020-11-23
Hi!

Sorry I've taken so long to see this.  It's a great change and has been mentioned in various issues.  I'm inviting you to be a contributor as I no longer actively maintain this project.

Cheers,

Steve