@cabauman 2020-11-23
Not a problem at all @SteveLillis . And I'd be happy to!